Refactored publisher list and publisher add into modular components that can be moved independently. Bug remains that after selecting a publisher, the menu doesn't reset.
