+++ app/components/snapshot.js
@@ -2,7 +2,8 @@ import React from 'react'
 import Fluxxor from 'fluxxor'
 import ActiveKeywordList from './activekeywordlist'
 import AddKeyword from './addkeyword'
-import PublisherList from './publisherlist'
+import ActivePublisherList from './activepublisherlist'
+import AddPublisher from './addpublisher'
 import Slider from './slider'
 import { Button } from 'react-bootstrap'
 
@@ -54,13 +55,11 @@ export default React.createClass({
           <i onClick={this.toggleHidden} className="fa fa-2x fa-cog chart-options"></i>
           <div className={(this.props.chartParams.hiddenSettings ? 'hidden ' : '') + "chart-menu"}>
             <h5>Keywords</h5>
-              <AddKeyword chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
-              <ActiveKeywordList chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
-            <PublisherList chartID={this.props.chartParams.chartID}
-                           className={'publisher-list'}
-                           list={this.props.publisherList}
-                           activelist={this.props.chartParams.publishers}
-                           legend={true}/>
+            <AddKeyword chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
+            <ActiveKeywordList chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
+            <h5>Publishers</h5>
+            <ActivePublisherList chartID={this.props.chartParams.chartID} className={'publisher-list'} list={this.props.publisherList} activelist={this.props.chartParams.publishers} legend={true} />
+            <AddPublisher chartID={this.props.chartParams.chartID} list={this.props.publisherList} activelist={this.props.chartParams.publishers} />
             <Button className="delete" bsStyle="danger">Delete Chart</Button>
           </div>
         </div>
